Refactoring model tests
